refactor: admin認証・dashboard・useChatSessionの3大技術的負債を解消
- withAdmin() HOFで5つのadmin APIルートの認証チェック重複を統合
- dashboard/page.tsx(553行)を6コンポーネント+2フックに分割
- useChatSession.ts(688行)を4サブフック+ユーティリティに分割
- 未使用のHistorySection.tsxを削除
- StatsSection.tsxをprops駆動に書き換え
- 機能・UIに変更なし、ビルド・Lint検証済み
