+++ src/app/api/admin/api-logs/route.ts
@@ -1,14 +1,8 @@
 import { NextResponse } from "next/server";
-import { auth } from "@/auth";
 import { createAdminClient } from "@/lib/supabase";
+import { withAdmin } from "@/lib/admin-auth";
 
-export async function GET() {
-  const session = await auth();
-  const role = session?.user?.role;
-  if (role !== "admin") {
-    return NextResponse.json({ error: "権限がありません" }, { status: 403 });
-  }
-
+export const GET = withAdmin(async () => {
   const supabase = createAdminClient();
 
   const { data, error } = await supabase
@@ -38,4 +32,4 @@ export async function GET() {
   }));
 
   return NextResponse.json({ logs });
-}
+});
